+++ src/mongo/dbtests/repltests.cpp
@@ -64,6 +64,31 @@ using std::string;
 using std::stringstream;
 using std::vector;
 
+/**
+ * Creates an OplogEntry with given parameters and preset defaults for this test suite.
+ */
+repl::OplogEntry makeOplogEntry(repl::OpTime opTime,
+                                repl::OpTypeEnum opType,
+                                NamespaceString nss,
+                                BSONObj object,
+                                boost::optional<BSONObj> object2) {
+    return repl::OplogEntry(opTime,                     // optime
+                            0,                          // hash
+                            opType,                     // opType
+                            nss,                        // namespace
+                            boost::none,                // uuid
+                            boost::none,                // fromMigrate
+                            OplogEntry::kOplogVersion,  // version
+                            object,                     // o
+                            object2,                    // o2
+                            {},                         // sessionInfo
+                            boost::none,                // wall clock time
+                            boost::none,                // statement id
+                            boost::none,   // optime of previous write within same transaction
+                            boost::none,   // pre-image optime
+                            boost::none);  // post-image optime
+}
+
 BSONObj f(const char* s) {
     return fromjson(s);
 }
@@ -78,6 +103,10 @@ protected:
 public:
     Base()
         : _client(&_opCtx), _defaultReplSettings(getGlobalReplicationCoordinator()->getSettings()) {
+        // Replication is not supported by mobile SE.
+        if (mongo::storageGlobalParams.engine == "mobile") {
+            return;
+        }
         ReplSettings replSettings;
         replSettings.setOplogSizeBytes(10 * 1024 * 1024);
         replSettings.setMaster(true);
@@ -91,7 +120,7 @@ public:
 
         getGlobalServiceContext()->setOpObserver(stdx::make_unique<OpObserverImpl>());
 
-        setOplogCollectionName();
+        setOplogCollectionName(getGlobalServiceContext());
         createOplog(&_opCtx);
 
         OldClientWriteContext ctx(&_opCtx, ns());
@@ -106,6 +135,10 @@ public:
         wuow.commit();
     }
     ~Base() {
+        // Replication is not supported by mobile SE.
+        if (mongo::storageGlobalParams.engine == "mobile") {
+            return;
+        }
         try {
             deleteAll(ns());
             deleteAll(cllNS());
@@ -274,6 +307,10 @@ protected:
 class LogBasic : public Base {
 public:
     void run() {
+        // Replication is not supported by mobile SE.
+        if (mongo::storageGlobalParams.engine == "mobile") {
+            return;
+        }
         ASSERT_EQUALS(2, opCount());
         _client.insert(ns(), fromjson("{\"a\":\"b\"}"));
         ASSERT_EQUALS(3, opCount());
@@ -286,6 +323,10 @@ class Base : public ReplTests::Base {
 public:
     virtual ~Base() {}
     void run() {
+        // Replication is not supported by mobile SE.
+        if (mongo::storageGlobalParams.engine == "mobile") {
+            return;
+        }
         reset();
         doIt();
         int nOps = opCount();
@@ -1267,6 +1308,10 @@ public:
 class DeleteOpIsIdBased : public Base {
 public:
     void run() {
+        // Replication is not supported by mobile SE.
+        if (mongo::storageGlobalParams.engine == "mobile") {
+            return;
+        }
         insert(BSON("_id" << 0 << "a" << 10));
         insert(BSON("_id" << 1 << "a" << 11));
         insert(BSON("_id" << 3 << "a" << 10));
@@ -1286,6 +1331,10 @@ public:
 class DatabaseIgnorerBasic {
 public:
     void run() {
+        // Replication is not supported by mobile SE.
+        if (mongo::storageGlobalParams.engine == "mobile") {
+            return;
+        }
         DatabaseIgnorer d;
         ASSERT(!d.ignoreAt("a", Timestamp(4, 0)));
         d.doIgnoreUntilAfter("a", Timestamp(5, 0));
@@ -1302,6 +1351,10 @@ public:
 class DatabaseIgnorerUpdate {
 public:
     void run() {
+        // Replication is not supported by mobile SE.
+        if (mongo::storageGlobalParams.engine == "mobile") {
+            return;
+        }
         DatabaseIgnorer d;
         d.doIgnoreUntilAfter("a", Timestamp(5, 0));
         d.doIgnoreUntilAfter("a", Timestamp(6, 0));
@@ -1324,7 +1377,7 @@ public:
     bool returnEmpty;
     SyncTest() : SyncTail(nullptr, SyncTail::MultiSyncApplyFunc()), returnEmpty(false) {}
     virtual ~SyncTest() {}
-    virtual BSONObj getMissingDoc(OperationContext* opCtx, const BSONObj& o) {
+    BSONObj getMissingDoc(OperationContext* opCtx, const OplogEntry& oplogEntry) override {
         if (returnEmpty) {
             BSONObj o;
             return o;
@@ -1339,14 +1392,20 @@ public:
 class FetchAndInsertMissingDocument : public Base {
 public:
     void run() {
+        // Replication is not supported by mobile SE.
+        if (mongo::storageGlobalParams.engine == "mobile") {
+            return;
+        }
         bool threw = false;
-        BSONObj o = BSON("ns" << ns() << "o" << BSON("foo"
-                                                     << "bar")
-                              << "o2"
-                              << BSON("_id"
-                                      << "in oplog"
-                                      << "foo"
-                                      << "bar"));
+        auto oplogEntry = makeOplogEntry(OpTime(Timestamp(100, 1), 1LL),  // optime
+                                         OpTypeEnum::kUpdate,             // op type
+                                         NamespaceString(ns()),           // namespace
+                                         BSON("foo"
+                                              << "bar"),  // o
+                                         BSON("_id"
+                                              << "in oplog"
+                                              << "foo"
+                                              << "bar"));  // o2
 
         Lock::GlobalWrite lk(&_opCtx);
 
@@ -1356,7 +1415,7 @@ public:
             badSource.setHostname("localhost:123");
 
             OldClientContext ctx(&_opCtx, ns());
-            badSource.getMissingDoc(&_opCtx, o);
+            badSource.getMissingDoc(&_opCtx, oplogEntry);
         } catch (DBException&) {
             threw = true;
         }
@@ -1364,7 +1423,7 @@ public:
 
         // now this should succeed
         SyncTest t;
-        verify(t.fetchAndInsertMissingDocument(&_opCtx, o));
+        verify(t.fetchAndInsertMissingDocument(&_opCtx, oplogEntry));
         verify(!_client
                     .findOne(ns(),
                              BSON("_id"
@@ -1373,7 +1432,7 @@ public:
 
         // force it not to find an obj
         t.returnEmpty = true;
-        verify(!t.fetchAndInsertMissingDocument(&_opCtx, o));
+        verify(!t.fetchAndInsertMissingDocument(&_opCtx, oplogEntry));
     }
 };
